1. adjust wandb watch to log 'all' 2. disable test result summary 3. finalize configuration 4. force Tensor().float() for CLIP's data transformation.  5. redirect ODDataModule test_dataloader pipeline to include validation datasets. 6. lower learning rate.

+++ src/clip/clip.py
@@ -83,7 +83,7 @@ def _to_tensor(image):
     if (isinstance(image, torch.Tensor)):
         if (image.max() > 1):
             image = image / 255
-        return image
+        return image.float()
     else:
         return ToTensor()(image)
